Fix swagger codegen-related bugs

+++ app/src/main/java/hu/bme/aut/stewe/rebrickableclient/network/swagger/api/LegoApi.kt
@@ -17,7 +17,7 @@ interface LegoApi {
 
     @GET("lego/themes/{id}/")
     fun legoThemesRead(
-            @Path("id") id: Int?
+            @Path("id") id: Long?
     ): Call<Theme>
